Move core functionality to module
Move the core of the bin/viirs_dr_runner.py functionality to a module.
Rather than defining globals in a `__main__` block, extract those
options necessary and pass those to the various functions as needed.

With this move, it will be possible to write unit tests for the
functionality included here.  When those unit tests are written, it will
be possible to extend or refactor the functionality.
